project/requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
python-telegram-bot-calendar==1.0.5
peewee==3.15.4